pyproject.toml
[project]
name = "translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.43.0",
    "docling>=2.10.0",
    "ipykernel>=6.29.5",
    "openai>=1.56.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "sounddevice>=0.5.1",
    "textual>=3.1.0",
]
